+++ src/Krystal/I18n/CHANGELOG.md
@@ -4,6 +4,7 @@ CHANGELOG
 1.3
 ---
 
+ * Added additional feature to filter translations by modules
  * The main `translate()` method no longer throws an exception if first argument isn't scalar. It ignores this and returns nothing instead
  * Since now `translateArray()` can translate two dimensional arrays
  * Fixed issue with scalar placeholders